Change 'model' to 'recipe'
Where we refer to 'model' right now it actually refers to a 'recipe'.
That is a combination of prompt generator and SAM model.

Mobile SAM v2 - is a recipe that uses a YOLO model to find bounding
boxes, uses the bounding box as a prompt then uses an 'efficient SAM'
model to generateo segmentations.

Sam Automatic Mask Generator - is a recipe that uses a grid of points as
the prompts, then uses 'vitb' model to segment.

Currently we only allow the user to choose between the above two recipes,
but in the near future I anticipate the GUI will become more flexible so
user can customize the prompt generator and the SAM model to create
custom recipes.